[pruner] generalize bias hook for conv2d
The bias hook is a forward hook that is part of the pruning parametrization; it is attached after the activation reconstruction forward hook, so adding the bias occurs after zeros are reinserted to the pruned activation.

This diff/PR amends the bias hook to work for Conv2d layers, in addition to Linear layers. The reshaping of the ._bias parameter ensures that it is added to the right dimension of the output.

Differential Revision: [D29979571](https://our.internmc.facebook.com/intern/diff/D29979571/)

ghstack-source-id: 134645733
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62430