Error on print statements in tests
Calls to `print` and other output during tests clogs up CI output and
makes it hard to follow. This adds a check for anything written to
stdout or stderr and raises an error if anything is there.

The codebase is not clean with this right now so there is an allowlist
that can be added to test classes: `_ignore_error_on_print_allowlist`.
It can also be enabled locally (it's off by default locally since people
commonly print in tests to debug / develop), with
`PYTORCH_ERROR_ON_TEST_PRINT=1`

ghstack-source-id: 03592ad8da8cfdd6ee493c8a3126578ec5c08b39
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62637