ns for fx: add ref_node_target_type
Summary:

Adds a `ref_node_target_type` field to hold the string type
of the base node. This is needed because in some cases
the previous node does not match ref_node (if we have observers,
or if we are logging inputs), and it is useful to know the type
of ref_node.

Test Plan:

```
python test/test_quantization.py TestFXNumericSuiteCoreAPIs
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: f7631417273b9a2d6de964b880413678d451101a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62685